requirements.txt
fastapi
uvicorn
pydantic
torch
transformers
langchain_community
langchain
accelerate
motor
faiss-cpu
python-dotenv
sentence-transformers
nest_asyncio


